+++ lib/database_cleaner/cleaners.rb
